+++ example/src/App.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import { UseEventListener } from './UseEventListener'
 import { UseInput } from './UseInput'
 import { UseLocalStorage } from './UseLocalStorage'
 
@@ -7,6 +8,7 @@ const App = () => {
     <>
       <UseInput />
       <UseLocalStorage />
+      <UseEventListener />
     </>
   )
 }